update screen navigation: add average screen
