+++ app/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_runtime_version"
     implementation "androidx.activity:activity-compose:$activity_compose_version"
+    implementation "androidx.navigation:navigation-compose:$compose_nav_version"
 
     implementation "androidx.hilt:hilt-navigation-compose:$hilt_navigation_version"
     implementation "com.google.dagger:hilt-android:$hilt_version"
