+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="com.ekyrizky.stonkscalculator">
 
     <application
+        android:name=".StonksApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -14,7 +15,7 @@
         android:theme="@style/Theme.StonksCalculator"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.StonksCalculator">
