+++ app/src/main/java/com/ekyrizky/stonkscalculator/StonksApplication.kt
@@ -1,0 +1,7 @@
+package com.ekyrizky.stonkscalculator
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class StonksApplication : Application()
