+++ app/src/main/java/com/ekyrizky/stonkscalculator/common/Constant.kt
@@ -1,0 +1,6 @@
+package com.ekyrizky.stonkscalculator.common
+
+object ScreenType {
+    const val profit = "profit_screen"
+    const val average = "average_screen"
+}
