+++ app/src/main/java/com/ekyrizky/stonkscalculator/presentation/average/AverageScreen.kt
@@ -1,0 +1,9 @@
+package com.ekyrizky.stonkscalculator.presentation.average
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AverageScreen() {
+    Text(text = "AVERAGE SCREEN")
+}
