+++ app/src/main/java/com/ekyrizky/stonkscalculator/presentation/utils/Screen.kt
@@ -1,0 +1,8 @@
+package com.ekyrizky.stonkscalculator.presentation.utils
+
+import com.ekyrizky.stonkscalculator.common.ScreenType
+
+sealed class Screen(val route: String) {
+    object ProfitScreen : Screen(ScreenType.profit)
+    object AverageScreen : Screen(ScreenType.average)
+}
