+++ build.gradle
@@ -5,6 +5,7 @@ buildscript {
         core_version = '1.8.0'
         compose_version = '1.2.1'
         compose_compiler_version = '1.3.1'
+        compose_nav_version = '2.5.3'
         activity_compose_version = '1.5.1'
         lifecycle_runtime_version = '2.5.1'
